fix(parser): treat a STEP comment as trivia when decoding an attribute value (#3673)
PR #3675 fixed the entity SCANNER: a `/* ... */` comment inside a record
no longer truncates or drops the record's byte span. It deliberately left
the attribute-DECODE layer comment-blind, documented in #3673 — with a
correct span, decoding still read the comment's own text into the value
it precedes (`/* rev; b */ $` decoded as the string `"/* rev; b */ $"`,
not `null`), and a comment before an unset `$` made
`has_non_null_attribute` report it as set.

Fixes all three decoders #3673 named:
- TypeScript `EntityExtractor` (attribute splitter and nested-list
  parser), reusing `StepTextScan` from step-lexing.ts.
- Rust's nom tokenizer: `ws` now shares `skip_step_trivia` with the
  scanners.
- Rust's `EntityScanner::has_non_null_attribute`: the leading-whitespace
  probe and the comma-separator walk both skip a comment as a region.

tokenizer.rs's inline test module is split into tokenizer_tests.rs
(mirroring scanner.rs/scanner_tests.rs) to stay inside its module-size
ratchet budget after the new tests.

Composition unchanged in both directions: a `/*` inside a string literal
is still text; a `'`, `;`, `(` or `,` inside a comment is still text, not
structure. A comment-free record decodes byte-identically to before.
