+++ rust/core/src/parser/tokenizer.rs
@@ -8,7 +8,7 @@
 
 use nom::{
     branch::alt,
-    bytes::complete::{take_while, take_while1},
+    bytes::complete::take_while1,
     character::complete::{char, digit1, one_of},
     combinator::{map, map_res, opt, recognize},
     multi::separated_list0,
@@ -175,9 +175,20 @@ fn typed_value_at_depth(input: &[u8], depth: u32) -> IResult<&[u8], Token<'_>> {
     )(input)
 }
 
-/// Skip whitespace
+/// Skip whitespace AND `/* ... */` comments -- 10303-21 allows a comment
+/// anywhere whitespace is allowed, including between the attributes this
+/// tokenizer decodes. Without this, `/* rev; b */ $` was not trivia: no `alt`
+/// branch in `token_at_depth` starts with `/`, so a comment before a value
+/// made the whole entity fail to parse rather than reading the value clean.
+///
+/// Reuses `skip_step_trivia`, the matched pair of the scanners' trivia skip
+/// (`step-lexing.ts`'s `skipTrivia` on the TS side). An unterminated comment
+/// consumes to end of input rather than failing this always-succeeds
+/// combinator; the `alt` that follows then has nothing left to match and the
+/// entity fails to parse, which is the existing behaviour for corrupt input.
 fn ws(input: &[u8]) -> IResult<&[u8], ()> {
-    map(take_while(|c: u8| c.is_ascii_whitespace()), |_| ())(input)
+    let end = super::lexical::skip_step_trivia(input, 0).unwrap_or(input.len());
+    Ok((&input[end..], ()))
 }
 
 /// Parse a token with optional surrounding whitespace
@@ -281,183 +292,5 @@ where
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// Table-driven basic token parsing: (parser, input, expected token).
-    #[test]
-    #[allow(clippy::approx_constant)]
-    fn test_basic_tokens() {
-        type Parser = for<'a> fn(&'a [u8]) -> IResult<&'a [u8], Token<'a>>;
-        let cases: &[(Parser, &[u8], Token)] = &[
-            (entity_ref, b"#123", Token::EntityRef(123)),
-            (entity_ref, b"#0", Token::EntityRef(0)),
-            (string_literal, b"'hello'", Token::String(b"hello")),
-            (
-                string_literal,
-                b"'with spaces'",
-                Token::String(b"with spaces"),
-            ),
-            (integer, b"42", Token::Integer(42)),
-            (integer, b"-42", Token::Integer(-42)),
-            (integer, b"0", Token::Integer(0)),
-            (float, b"3.14", Token::Float(3.14)),
-            (float, b"-3.14", Token::Float(-3.14)),
-            (float, b"1.5E-10", Token::Float(1.5e-10)),
-            (enum_value, b".TRUE.", Token::Enum(b"TRUE")),
-            (enum_value, b".FALSE.", Token::Enum(b"FALSE")),
-            (enum_value, b".ELEMENT.", Token::Enum(b"ELEMENT")),
-        ];
-        for (parse, input, expected) in cases {
-            assert_eq!(
-                parse(input),
-                Ok((&b""[..], expected.clone())),
-                "tokenizing {input:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn test_list() {
-        let result = list(b"(1,2,3)");
-        assert!(result.is_ok());
-        let (_, token) = result.unwrap();
-        match token {
-            Token::List(items) => {
-                assert_eq!(items.len(), 3);
-                assert_eq!(items[0], Token::Integer(1));
-                assert_eq!(items[1], Token::Integer(2));
-                assert_eq!(items[2], Token::Integer(3));
-            }
-            _ => panic!("Expected List token"),
-        }
-    }
-
-    #[test]
-    fn test_nested_list() {
-        let result = list(b"(1,(2,3),4)");
-        assert!(result.is_ok());
-        let (_, token) = result.unwrap();
-        match token {
-            Token::List(items) => {
-                assert_eq!(items.len(), 3);
-                assert_eq!(items[0], Token::Integer(1));
-                match &items[1] {
-                    Token::List(inner) => {
-                        assert_eq!(inner.len(), 2);
-                        assert_eq!(inner[0], Token::Integer(2));
-                        assert_eq!(inner[1], Token::Integer(3));
-                    }
-                    _ => panic!("Expected nested List"),
-                }
-                assert_eq!(items[2], Token::Integer(4));
-            }
-            _ => panic!("Expected List token"),
-        }
-    }
-
-    #[test]
-    fn test_parse_entity() {
-        let input = "#123=IFCWALL('guid','owner',$,$,'name',$,$,$);";
-        let result = parse_entity(input);
-        assert!(result.is_ok());
-        let (id, ifc_type, args) = result.unwrap();
-        assert_eq!(id, 123);
-        assert_eq!(ifc_type, IfcType::IfcWall);
-        assert_eq!(args.len(), 8);
-    }
-
-    #[test]
-    fn test_parse_entity_with_nested_list() {
-        // First test: simple list (should work)
-        let simple = "(0.,0.,1.)";
-        println!("Testing simple list: {}", simple);
-        let simple_result = list(simple.as_bytes());
-        println!("Simple list result: {:?}", simple_result);
-
-        // Second test: nested in entity (what's failing)
-        let input = "#9=IFCDIRECTION((0.,0.,1.));";
-        println!("\nTesting full entity: {}", input);
-        let result = parse_entity(input);
-
-        if let Err(ref e) = result {
-            println!("Parse error: {:?}", e);
-
-            // Try parsing just the arguments part
-            println!("\nTrying to parse just arguments: ((0.,0.,1.))");
-            let args_input = "((0.,0.,1.))";
-            let args_result = list(args_input.as_bytes());
-            println!("Args list result: {:?}", args_result);
-        }
-
-        assert!(result.is_ok(), "Failed to parse: {:?}", result);
-        let (id, _ifc_type, args) = result.unwrap();
-        assert_eq!(id, 9);
-        assert_eq!(args.len(), 1);
-        // First arg should be a list containing 3 floats
-        if let Token::List(inner) = &args[0] {
-            assert_eq!(inner.len(), 3);
-        } else {
-            panic!("Expected Token::List, got {:?}", args[0]);
-        }
-    }
-
-    /// Deeply nested list arguments must return an error rather than
-    /// recursing through the stack until it overflows.
-    #[test]
-    fn test_parse_entity_rejects_excessive_nesting() {
-        let n = (MAX_NESTING_DEPTH as usize) + 64;
-        let mut s = String::from("#1=IFCWALL(");
-        for _ in 0..n {
-            s.push('(');
-        }
-        s.push('1');
-        for _ in 0..n {
-            s.push(')');
-        }
-        s.push_str(");");
-        // Must not panic / overflow; must return Err.
-        assert!(parse_entity(&s).is_err());
-    }
-
-    /// Moderate nesting still parses successfully.
-    #[test]
-    fn test_parse_entity_accepts_moderate_nesting() {
-        let n = 32;
-        let mut s = String::from("#1=IFCWALL(");
-        for _ in 0..n {
-            s.push('(');
-        }
-        s.push('1');
-        for _ in 0..n {
-            s.push(')');
-        }
-        s.push_str(");");
-        assert!(parse_entity(&s).is_ok());
-    }
-
-    fn nested(n: usize) -> String {
-        let mut s = String::from("#1=IFCWALL(");
-        for _ in 0..n {
-            s.push('(');
-        }
-        s.push('1');
-        for _ in 0..n {
-            s.push(')');
-        }
-        s.push_str(");");
-        s
-    }
-
-    /// Boundary: parsing succeeds exactly at MAX_NESTING_DEPTH.
-    #[test]
-    fn test_parse_entity_accepts_exactly_max_nesting() {
-        assert!(parse_entity(&nested(MAX_NESTING_DEPTH as usize)).is_ok());
-    }
-
-    /// Boundary: parsing fails at MAX_NESTING_DEPTH + 1.
-    #[test]
-    fn test_parse_entity_rejects_one_over_max_nesting() {
-        assert!(parse_entity(&nested(MAX_NESTING_DEPTH as usize + 1)).is_err());
-    }
-}
+#[path = "tokenizer_tests.rs"]
+mod tokenizer_tests;
